fix for graph mode

+++ logbesselk/math.py
@@ -63,11 +63,11 @@ def log_cosh(x):
 
 def sinc(x):
     pix = np.pi * x
-    return tf.where(tf.equal(x, 0.), 1., sin(pix) / pix)
+    return tf.where(tf.equal(x, 0.), tf.cast(1., x.dtype), sin(pix) / pix)
 
 
 def sinhc(x):
-    return tf.where(tf.equal(x, 0.), 1., sinh(x) / x)
+    return tf.where(tf.equal(x, 0.), tf.cast(1., x.dtype), sinh(x) / x)
 
 
 def log_add_exp(x, y, sign=None):
